+++ src/providers/ts.ts
@@ -50,7 +50,11 @@ function buildLexer() {
             a
               .from(/`(?:\\.|[^\\`$])*(?:\$\{|`|$)/)
               // reject if ends with '${'
-              .reject(({ output }) => output.content.endsWith("${")),
+              .reject(({ output }) => output.content.endsWith("${"))
+              .data(({ output }) => ({
+                // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+                unclosed: !output.content.split(/\\./).at(-1)!.endsWith("`"),
+              })),
         }
       )
       .define({
@@ -147,7 +151,9 @@ export class TsStringParser implements IStringParser {
           return;
         }
 
-        return evalTsString(token.content);
+        return Lexer.javascript.evalString(
+          token.data.unclosed ? token.content + token.content[0] : token.content
+        );
       }
 
       // if the hover is in a template string, set targetTempStrIndex
@@ -177,7 +183,7 @@ export class TsStringParser implements IStringParser {
         if (targetTempStrIndex === tempStrStack.length) {
           // got the target template string, calculate string value
           const quoted = tokens.map((t) => t.content).join("...");
-          return evalTsString(quoted);
+          return Lexer.javascript.evalString(quoted); // TODO: handle unclosed
         }
       }
 
@@ -195,55 +201,4 @@ export class TsStringParser implements IStringParser {
       // else, got token but not string, continue
     }
   }
-}
-
-// TODO: make this a Lexer utils in retsac
-function evalTsString(quoted: string) {
-  // remove quotes
-  const quote = quoted[0];
-  const unquoted = quoted.slice(
-    // remove the first quote
-    1,
-    // the string might be un-closed, so the last char might not be the quote.
-    quoted.at(-1) === quote ? -1 : undefined
-  );
-
-  // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Lexical_grammar#literals
-  // IMPORTANT! all escaped chars should be searched simultaneously!
-  // e.g. you should NOT use `unquoted.replace(/\\\\/g, "\\").replace(/\\'/g, "'")...`
-  return unquoted.replace(
-    /(\\0|\\'|\\"|\\n|\\\\|\\r|\\v|\\t|\\b|\\f|\\\n|\\`|\\x([0-9a-fA-F]{2})|\\u([0-9a-fA-F]{4}))/g,
-    (match) => {
-      if (match === `\\0`) {
-        return "\0";
-      } else if (match === `\\'`) {
-        return "'";
-      } else if (match === `\\"`) {
-        return '"';
-      } else if (match === `\\n`) {
-        return "\n";
-      } else if (match === `\\\\`) {
-        return "\\";
-      } else if (match === `\\r`) {
-        return "\r";
-      } else if (match === `\\v`) {
-        return "\v";
-      } else if (match === `\\t`) {
-        return "\t";
-      } else if (match === `\\b`) {
-        return "\b";
-      } else if (match === `\\f`) {
-        return "\f";
-      } else if (match === `\\\n`) {
-        return "";
-      } else if (match === "\\`") {
-        return "`";
-      } else if (match.startsWith("\\x")) {
-        return String.fromCharCode(parseInt(match.slice(2), 16));
-      } else {
-        // match.startsWith("\\u")
-        return String.fromCharCode(parseInt(match.slice(2), 16));
-      }
-    }
-  );
 }