+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v0.1.7
 
 - Feat(json/jsonc): support unclosed string.
+- Fix(ts): support line continuation. [#1](https://github.com/DiscreteTom/string-hover/issues/1)
+- Fix(ts): support unicode code point escaped. [#2](https://github.com/DiscreteTom/string-hover/issues/2)
 
 ## v0.1.6