+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## v0.1.8
+
+- Feat(json): error handling for invalid escape sequence with retsac v0.15.0 built-in json utils.
+
 ## v0.1.7
 
 - Feat(json/jsonc/ts): support unclosed string.